PA-34598 Make each package independently releaseable (#53)
