+++ inscacheable/go.mod
@@ -1,0 +1,15 @@
+module github.com/useinsider/go-pkg/inscacheable
+
+go 1.19
+
+require (
+	github.com/jellydator/ttlcache/v3 v3.0.0
+	github.com/stretchr/testify v1.8.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
