+++ insgorm/go.mod
@@ -1,0 +1,15 @@
+module github.com/useinsider/go-pkg/insgorm
+
+go 1.19
+
+require (
+	github.com/DATA-DOG/go-sqlmock v1.5.0
+	gorm.io/driver/mysql v1.3.4
+	gorm.io/gorm v1.23.7
+)
+
+require (
+	github.com/go-sql-driver/mysql v1.7.0 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+)
