+++ insgorm/gorm_mock.go
@@ -1,0 +1,21 @@
+package insgorm
+
+import (
+	"github.com/DATA-DOG/go-sqlmock"
+	"gorm.io/driver/mysql"
+	"gorm.io/gorm"
+)
+
+func MockGorm() (*gorm.DB, sqlmock.Sqlmock) {
+	sqlDB, mock, err := sqlmock.New(sqlmock.QueryMatcherOption(sqlmock.QueryMatcherEqual))
+	if err != nil {
+		panic(err)
+	}
+
+	gormDB, err := gorm.Open(mysql.New(mysql.Config{Conn: sqlDB, SkipInitializeWithVersion: true}), &gorm.Config{})
+	if err != nil {
+		panic(err)
+	}
+
+	return gormDB, mock
+}
