+++ inskinesis/go.mod
@@ -1,0 +1,17 @@
+module github.com/useinsider/go-pkg/inskinesis
+
+go 1.19
+
+require (
+	github.com/aws/aws-sdk-go v1.44.3
+	github.com/google/uuid v1.3.1
+	github.com/stretchr/testify v1.8.1
+	go.uber.org/mock v0.3.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
