+++ inslogger/go.mod
@@ -1,0 +1,7 @@
+module github.com/useinsider/go-pkg/inslogger
+
+go 1.19
+
+require go.uber.org/zap v1.26.0
+
+require go.uber.org/multierr v1.10.0 // indirect
