+++ insredis/go.mod
@@ -1,0 +1,13 @@
+module github.com/useinsider/go-pkg/insredis
+
+go 1.23.0
+
+require (
+	github.com/go-redis/redis v6.15.9+incompatible
+	github.com/golang/mock v1.6.0
+)
+
+require (
+	github.com/onsi/ginkgo v1.16.5 // indirect
+	github.com/onsi/gomega v1.38.3 // indirect
+)
