+++ insrequester/go.mod
@@ -1,0 +1,23 @@
+module github.com/useinsider/go-pkg/insrequester
+
+go 1.19
+
+require (
+	github.com/golang/mock v1.6.0
+	github.com/pkg/errors v0.9.1
+	github.com/slok/goresilience v0.2.0
+	github.com/stretchr/testify v1.8.1
+)
+
+require (
+	github.com/beorn7/perks v0.0.0-20180321164747-3a771d992973 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/golang/protobuf v1.2.0 // indirect
+	github.com/matttproud/golang_protobuf_extensions v1.0.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/prometheus/client_golang v0.9.2 // indirect
+	github.com/prometheus/client_model v0.0.0-20180712105110-5c3871d89910 // indirect
+	github.com/prometheus/common v0.0.0-20181126121408-4724e9255275 // indirect
+	github.com/prometheus/procfs v0.0.0-20181204211112-1dc9a6cbc91a // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
