+++ inssentry/go.mod
@@ -1,0 +1,7 @@
+module github.com/useinsider/go-pkg/inssentry
+
+go 1.19
+
+require github.com/getsentry/sentry-go v0.13.0
+
+require golang.org/x/sys v0.6.0 // indirect
