+++ inssimpleroute/go.mod
@@ -1,0 +1,3 @@
+module github.com/useinsider/go-pkg/inssimpleroute
+
+go 1.19
