+++ inssql/go.mod
@@ -1,0 +1,5 @@
+module github.com/useinsider/go-pkg/inssql
+
+go 1.19
+
+require github.com/DATA-DOG/go-sqlmock v1.5.0
