+++ inssql/sql_mock.go
@@ -3,10 +3,9 @@ package inssql
 import (
 	"database/sql"
 	"database/sql/driver"
-	"github.com/DATA-DOG/go-sqlmock"
-	"gorm.io/driver/mysql"
-	"gorm.io/gorm"
 	"time"
+
+	"github.com/DATA-DOG/go-sqlmock"
 )
 
 func MockSql() (*sql.DB, sqlmock.Sqlmock) {
@@ -18,23 +17,8 @@ func MockSql() (*sql.DB, sqlmock.Sqlmock) {
 	return sqlDB, mock
 }
 
-func MockGorm() (*gorm.DB, sqlmock.Sqlmock) {
-	sqlDB, mock, err := sqlmock.New(sqlmock.QueryMatcherOption(sqlmock.QueryMatcherEqual))
-	if err != nil {
-		panic(err)
-	}
-
-	gormDB, err := gorm.Open(mysql.New(mysql.Config{Conn: sqlDB, SkipInitializeWithVersion: true}), &gorm.Config{})
-	if err != nil {
-		panic(err)
-	}
-
-	return gormDB, mock
-}
-
 type AnyTime struct{}
 
-// Match satisfies sqlmock.Argument interface
 func (a AnyTime) Match(v driver.Value) bool {
 	_, ok := v.(time.Time)
 	return ok
