+++ insssm/go.mod
@@ -1,0 +1,16 @@
+module github.com/useinsider/go-pkg/insssm
+
+go 1.19
+
+require (
+	github.com/Jamil-Najafov/go-aws-ssm v0.9.0
+	github.com/useinsider/go-pkg/inscacheable v1.0.0
+)
+
+require (
+	github.com/aws/aws-sdk-go v1.44.3 // indirect
+	github.com/jellydator/ttlcache/v3 v3.0.0 // indirect
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
+	github.com/mitchellh/mapstructure v1.5.0 // indirect
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
+)
